+++ core/impl/src/main/java/com/blazebit/persistence/impl/AbstractFullQueryBuilder.java
@@ -447,7 +447,7 @@ public abstract class AbstractFullQueryBuilder<T, X extends FullQueryBuilder<T, 
         // Note: Identifier expressions are inner joined!
         List<ResolvedExpression> resolvedExpressions = new ArrayList<>(identifierExpressions == null ? 1 : identifierExpression.length() + 1);
         Expression expression = expressionFactory.createSimpleExpression(identifierExpression, false);
-        joinManager.implicitJoin(expression, true, false, null, null, JoinType.INNER, new HashSet<String>(), false, false, true, false, false, false);
+        joinManager.implicitJoin(expression, true, false, null, null, JoinType.INNER, null, new HashSet<String>(), false, false, true, false, false, false);
         StringBuilder sb = new StringBuilder();
 
         implicitJoinWhereClause();
@@ -469,7 +469,7 @@ public abstract class AbstractFullQueryBuilder<T, X extends FullQueryBuilder<T, 
         if (identifierExpressions != null) {
             for (String expressionString : identifierExpressions) {
                 expression = expressionFactory.createSimpleExpression(expressionString, false);
-                joinManager.implicitJoin(expression, true, false, null, null, JoinType.INNER, new HashSet<String>(), false, false, true, false, false, false);
+                joinManager.implicitJoin(expression, true, false, null, null, JoinType.INNER, null, new HashSet<String>(), false, false, true, false, false, false);
                 functionalDependencyAnalyzerVisitor.analyzeFormsUniqueTuple(expression);
                 if (functionalDependencyAnalyzerVisitor.getSplittedOffExpressions().isEmpty()) {
                     sb.setLength(0);
@@ -628,7 +628,7 @@ public abstract class AbstractFullQueryBuilder<T, X extends FullQueryBuilder<T, 
     public X fetch(String path) {
         prepareForModification(ClauseType.JOIN);
         verifyBuilderEnded();
-        joinManager.implicitJoin(expressionFactory.createPathExpression(path), true, true, null, null, null, new HashSet<String>(), false, false, true, false, true, false);
+        joinManager.implicitJoin(expressionFactory.createPathExpression(path), true, true, null, null, null, null, new HashSet<String>(), false, false, true, false, true, false);
         return (X) this;
     }
 
@@ -640,7 +640,7 @@ public abstract class AbstractFullQueryBuilder<T, X extends FullQueryBuilder<T, 
 
         HashSet<String> currentlyResolvingAliases = new HashSet<>();
         for (String path : paths) {
-            joinManager.implicitJoin(expressionFactory.createPathExpression(path), true, true, null, null, null, currentlyResolvingAliases, false, false, true, false, true, false);
+            joinManager.implicitJoin(expressionFactory.createPathExpression(path), true, true, null, null, null, null, currentlyResolvingAliases, false, false, true, false, true, false);
         }
 
         return (X) this;