+++ CHANGELOG.md
@@ -13,6 +13,7 @@ None
 ### Bug fixes
 
 * Fix support for de-referencing map key entity type expressions
+* Fix for absolute paths referring to dependent joins in ON clauses
 
 ### Backwards-incompatible changes